# openapi.yaml
openapi: 3.0.3
info:
  title: CodeLens API
  description: AI-powered code review platform API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "201": { description: User created, returns token + user }
        "409": { description: Email already registered }
  /login:
    post:
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: Returns token + user }
        "401": { description: Invalid credentials }
...
  /repo/{id}/rag/index:
    post:
      summary: Chunk and embed a repository for chat retrieval
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Indexing summary (files, chunks) }
  /repo/{id}/rag/chat:
    post:
      summary: Ask a question about the repository (RAG-grounded answer)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string }
      responses:
        "200": { description: Answer with cited source chunks }
  /account:
    delete:
      summary: Delete the authenticated user's account and all associated data
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Account deleted }